requirements.txt
# YT-DLP Web V2 - 轻量化依赖
# 核心框架
Flask>=3.1.1
Flask-CORS>=5.0.0

# 认证和安全
PyJWT>=2.10.1

# HTTP请求
requests>=2.32.0

# 配置管理
PyYAML>=6.0.2

# Telegram集成 (可选)
pyrogram>=2.1.0
TgCrypto>=1.3.0

# yt-dlp (核心功能)
yt-dlp>=2024.12.13

# 生产部署 (可选)
gunicorn>=23.0.0
python-dotenv>=1.0.1

# 开发工具 (可选)
# pytest>=8.0.0
# black>=24.0.0
# flake8>=7.0.0
